Add creation of crate engines from BeamNG mods in the crate tab
