+++ Cargo.toml
@@ -57,6 +57,7 @@ rfd = "0.12.0"
 winit = "0.28.1"
 itertools = "0.11.0"
 config = { version="0.13.1", features = ["toml"] }
+sanitize-filename = "0.5.0"
 
 
 [target.'cfg(windows)'.build-dependencies]
